+++ nussl/ml/train/trainer.py
@@ -13,9 +13,6 @@ import torch
 from torch import nn
 import numpy as np
 
-from nussl import datasets
-
-
 class ValidationEvents(EventEnum):
     """
     Events based on validation running
@@ -143,24 +140,6 @@ def create_train_and_validation_engines(train_func, val_func=None, device='cpu')
 
     return trainer, validator
 
-
-def _remove_cache_from_tfms(transforms):
-    transforms = copy.deepcopy(transforms)
-
-    if isinstance(transforms, datasets.transforms.Compose):
-        for t in transforms.transforms:
-            if isinstance(t, datasets.transforms.Cache):
-                transforms.transforms.remove(t)
-
-    return transforms
-
-
-def _prep_metadata(metadata):
-    metadata = copy.deepcopy(metadata)
-    metadata['transforms'] = _remove_cache_from_tfms(metadata['transforms'])
-    return metadata
-
-
 def add_validate_and_checkpoint(output_folder, model, optimizer, train_data, trainer,
                                 val_data=None, validator=None, save_by_epoch=None):
     """
@@ -233,45 +212,24 @@ def add_validate_and_checkpoint(output_folder, model, optimizer, train_data, tra
                 output_folder, 'checkpoints', 'best.model.pth'
             ))
 
-        metadata = {
-            'stft_params': train_data.stft_params,
-            'sample_rate': train_data.sample_rate,
-            'num_channels': train_data.num_channels,
-            'train_dataset': _prep_metadata(train_data.metadata),
-            'trainer.state_dict': {
-                'epoch': trainer.state.epoch,
-                'epoch_length': trainer.state.epoch_length,
-                'max_epochs': trainer.state.max_epochs,
-                'output': trainer.state.output,
-                'metrics': trainer.state.metrics,
-                'seed': trainer.state.seed,
-            },
-            'trainer.state.epoch_history': trainer.state.epoch_history,
-        }
-        try:
-            # Store metadata for validation set if it exists
-            metadata['val_dataset'] = _prep_metadata(val_data.metadata)
-        except:
-            pass
-
         if isinstance(model, nn.DataParallel):
             _model = model.module
         else:
             _model = model
 
-        _model.metadata.update(metadata)
-
         for _path in output_paths:
             os.makedirs(os.path.join(
                 output_folder, 'checkpoints'), exist_ok=True)
-            _model.save(_path)
+            _model.save(_path, train_data=train_data, val_data=val_data, 
+                        trainer=trainer)
             torch.save(optimizer.state_dict(),
                        _path.replace('model.pth', 'optimizer.pth'))
         
         if save_by_epoch is not None:
             if trainer.state.epoch % save_by_epoch == 0:
                 _path = output_paths[0].replace('latest', f'epoch{trainer.state.epoch}')
-                _model.save(_path)
+                _model.save(_path, train_data=train_data, val_data=val_data, 
+                            trainer=trainer)
 
         trainer.state.saved_model_path = output_paths[-1]
         trainer.state.output_folder = output_folder