+++ nussl/separation/base/deep_mixin.py
@@ -4,7 +4,6 @@ import json
 
 from ...ml import SeparationModel
 from ...datasets import transforms as tfm
-from ...core.migration import SafeModelLoader
 
 OMITTED_TRANSFORMS = (
     tfm.GetExcerpt,
@@ -31,16 +30,11 @@ class DeepMixin:
             metadata (dict): metadata associated with model, used for making
             the input data into the model.
         """
-        safe_loader = SafeModelLoader()
-        model_dict = safe_loader.load(model_path, 'cpu')
-        metadata = model_dict['metadata']
-
-        model = SeparationModel(metadata['config'])
-        model.load_state_dict(model_dict['state_dict'])
+        
         device = device if torch.cuda.is_available() else 'cpu'
 
         self.device = device
-
+        model, metadata = SeparationModel.load(model_path)
         model = model.to(device).eval()
         self.model = model
         self.config = metadata['config']
@@ -73,19 +67,35 @@ class DeepMixin:
                 transform = None
         return transform
 
-    def _get_input_data_for_model(self, extra_data=None):
+    def modify_input_data(self, data, **kwargs):
+        """Add or modify input data to dictionary before passing 
+        it to the model. By default this just updates the data
+        dictionary with what is needed, but can be overridden
+        by classes inheriting this method to modify the data
+        dictionary as needed.
+
+        Parameters
+        ----------
+        data : dict,
+            The data dictionary before this function is called.
+        kwargs : keyword arguments, optional
+            Data dictionary after this function is called, by default None
+        """
+        data.update(kwargs)
+        return data
+
+    def _get_input_data_for_model(self, **kwargs):
         """
         Sets up the audio signal with the appropriate STFT parameters and runs it
         through the transform found in the metadata.
 
         Args:
-            extra_data: A dictionary containing any additional data that will 
-              be merged with the output dictionary.
+            kwargs: Any additional data that will 
+              be merged with the input dictionary.
         
         Returns:
             dict: Data dictionary to pass into the model.
         """
-        extra_data = {} if extra_data is None else extra_data
         if self.metadata['sample_rate'] is not None:
             if self.audio_signal.sample_rate != self.metadata['sample_rate']:
                 self.audio_signal.resample(self.metadata['sample_rate'])
@@ -94,7 +104,6 @@ class DeepMixin:
         self.audio_signal.stft()
 
         data = {'mix': self.audio_signal}
-        data.update(extra_data)
         data = self.transform(data)
 
         for key in data:
@@ -103,6 +112,8 @@ class DeepMixin:
                 if self.metadata['num_channels'] == 1:
                     # then each channel is processed indep
                     data[key] = data[key].transpose(0, self.channel_dim)
+        
+        data = self.modify_input_data(data, **kwargs)       
         self.input_data = data
         return self.input_data