+++ tests/separation/test_deep.py
@@ -111,7 +111,7 @@ def test_deep_mixin(overfit_model):
     assert deep_mixin.audio_signal.sample_rate == deep_mixin.metadata['sample_rate']
 
     dummy_data = {'one_hot': np.random.rand(100)}
-    input_data = deep_mixin._get_input_data_for_model(dummy_data)
+    input_data = deep_mixin._get_input_data_for_model(**dummy_data)
 
     assert 'one_hot' in input_data